+++ lib/flash-rs/src/config/config_noclap.rs
@@ -2,8 +2,8 @@ use std::time::Duration;
 
 #[derive(Debug)]
 pub struct FlashConfig {
-    pub(crate) umem_id: u32,
-    pub(crate) nf_id: u32,
+    pub(crate) umem_id: u16,
+    pub(crate) nf_id: u16,
     pub(crate) smart_poll: bool,
     pub(crate) idle_timeout: Duration,
     pub(crate) idleness: f32,
